refactoring, improvements
https://github.com/metasfresh/metasfresh/issues/11362
